feat: add websocket stream
